+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [0.17.0] - 2019-09-20
+
 - Upgraded suggested version of python-click to >=7.0.
 - Fixed parsing of booleans from config files.
 - Fixed tolerance of booleans without values in config files (thanks to @Mno-hime for reporting).